Add an error check to static copying
--HG--
extra : convert_revision : svn%3Aa38d40e9-c014-0410-b785-c606c0c8e7de/satchmo/trunk%401133

+++ satchmo/shop/management/commands/satchmo_copy_static.py
@@ -9,5 +9,8 @@ class Command(NoArgsCommand):
         import satchmo
         static_src = os.path.join(satchmo.__path__[0],'static')
         static_dest = os.path.join(os.getcwd(), 'static')
-        shutil.copytree(static_src, static_dest)
-        print "Copied %s to %s" % (static_src, static_dest)
+        if os.path.exists(static_dest):
+            print "Static directory exists. You must manually copy the files you need."
+        else:
+            shutil.copytree(static_src, static_dest)
+            print "Copied %s to %s" % (static_src, static_dest)
